requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
Babel>=1.3
python-neutronclient>=2.3.10
-e git://git.openstack.org/openstack/neutron.git#egg=neutron
